lib: add more error context when loading fails
Previously the code path that read from `CertPaths` that configure a dir
from the environment lacked helpful context if the dir couldn't be read.
Similarly, the code path that read from a `CertPaths` file would add
helpful context if `rustls_pemfile` returned an error, but not if the
file open failed.

This commit centralizes the job of adding useful context into
`CertPaths.load()`. The `rustls_pemfile` specific part in the file path
processing can then be removed. This ensures for both dirs and files
that if the operation fails there's useful context as to why.