Prefer partial success to failure
If we read some certs from a file, but fail to read some from
a directory (or vv.) then return what we have.

This is good because it restores the crate's previous behaviour.
It also matches what (for example) golang crypto.x509 does.

This is bad because it hides a failure, which would be confusing
if a user relied on reading from _both_ a file and directory at
the same time.  Would someone do that?  The follow commit steps
towards ameliorating this, slightly.