tests: fix smoketests for 0.8 API
Several tests were broken with the API update. The `check_site()`
function was using `CertificateResult.unwrap()` and so panics if there
were any errors encountered. Some tests do cause errors, but also manage
to collect certs. For these tests we want more flexibility to
interrogate the returned errors, and then to use whatever roots we found
for a functional test with a `check_site()`-like fn.

Along the way I also relaxed the error message matching to only assert
some key details (whether it was a file or dir error, and what the
invalid file/dir was). The messages had shifted and the previous logic
was too fragile.